requirements.txt
streamlit==1.31.0
pandas==2.1.0
numpy==1.24.3
plotly==5.18.0
scipy==1.11.3
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
yfinance==0.2.28
requests==2.31.0
